Carrusel inmuebles implementado con éxito

+++ empresa/models.py
@@ -4,15 +4,16 @@ from django.db.models.signals import pre_save, pre_delete, post_save
 from django.dispatch import receiver
 from PIL import Image
 
-def nombrar_logo(instance, filename):#Función para guardado de imagen
+
+def nombrar_logo(instance, filename):  # Función para guardado de imagen
     ext = filename.split('.')[-1]
 
-    # Eliminar archivo antiguo si existe
-    existente = InformacionEmpresa.objects.get(pk=instance.pk)
-    if existente.logo:
-        ruta_anterior = existente.logo.path
-        if os.path.isfile(ruta_anterior):
-            os.remove(ruta_anterior)
+    if instance.id:  # Eliminar archivo antiguo si existe
+        existente = InformacionEmpresa.objects.get(pk=instance.pk)
+        if existente.logo:
+            ruta_anterior = existente.logo.path
+            if os.path.isfile(ruta_anterior):
+                os.remove(ruta_anterior)
 
     return f'Empresa/logo.{ext}'
 
@@ -29,7 +30,7 @@ class InformacionEmpresa(models.Model):
     correo_electronico = models.EmailField()
     title_boletin = models.CharField(max_length=255)
     descripcion_boletin = models.TextField()
-    logo = models.ImageField(upload_to = nombrar_logo, null=True, blank=True)
+    logo = models.ImageField(upload_to=nombrar_logo, null=True, blank=True)
     twitter = models.URLField(blank=True, null=True, unique=True)
     facebook = models.URLField(blank=True, null=True, unique=True)
     instagram = models.URLField(blank=True, null=True, unique=True)
@@ -37,11 +38,13 @@ class InformacionEmpresa(models.Model):
     msj_whatsapp = models.TextField()
     msj_asesores = models.TextField()
     autores = models.CharField(max_length=255)
+
     def __str__(self):
         return self.nombre_empresa.title()
 
 
-@receiver(pre_save, sender=InformacionEmpresa)#Señal que borra la imagen si está vacía
+# Señal que borra la imagen si está vacía
+@receiver(pre_save, sender=InformacionEmpresa)
 def borrar_logo_al_actualizar(sender, instance, **kwargs):
     # Verificar si la imagen ha cambiado
     if instance.id and not instance.logo:
@@ -50,8 +53,10 @@ def borrar_logo_al_actualizar(sender, instance, **kwargs):
             ruta_anterior = existente.logo.path
             if os.path.exists(ruta_anterior):
                 os.remove(ruta_anterior)
-                
-@receiver(pre_delete, sender=InformacionEmpresa)#Señal que borra el logo al ser eliminado su registro
+
+
+# Señal que borra el logo al ser eliminado su registro
+@receiver(pre_delete, sender=InformacionEmpresa)
 def borrar_logo_al_eliminar(sender, instance, **kwargs):
     # Eliminar archivo antiguo si existe
     if instance.logo:
@@ -59,7 +64,9 @@ def borrar_logo_al_eliminar(sender, instance, **kwargs):
         if os.path.exists(ruta_anterior):
             os.remove(ruta_anterior)
 
-@receiver(post_save, sender=InformacionEmpresa)#Señal para redimensionar imagen a un tamaño por defecto
+
+# Señal para redimensionar imagen a un tamaño por defecto
+@receiver(post_save, sender=InformacionEmpresa)
 def redimensionar_imagen(sender, instance, created, **kwargs):
     if instance.logo:
         try:
